update map api day 19

+++ FE/travel/src/api/Blog/index.js
@@ -4,19 +4,18 @@ import i18n from "../../i18n";
 
 export const createBlog = async (data) => {
     try {
-        const res = await axiosClient.post("/blogs", data);
+        const res = await axiosClient.post("/admin/blogs", data);
 
         toast.success(i18n.t("notify.create_success"));
         return res?.data;
     } catch (error) {
         toast.error(i18n.t("notify.create_fail"));
-        throw error;
     }
 };
 
 export const deleteBlog = async (id) => {
     try {
-        await axiosClient.delete(`/blogs/${id}`);
+        await axiosClient.delete(`/admin/blogs/${id}`);
 
         toast.success(i18n.t("notify.delete_success"));
     } catch (error) {
@@ -26,13 +25,26 @@ export const deleteBlog = async (id) => {
 
 export const getBlog = async (params) => axiosClient.get("/blogs", { params }).then((res) => res?.data);
 
-export const getMostReadBlog = async (params) => axiosClient.get("/blogs", { params }).then((res) => res?.data);
+export const getMostReadBlog = async (params) =>
+    axiosClient.get("/blogs/most-read", { params }).then((res) => res?.data);
 
-export const countBlog = async ({ id }) => axiosClient.get(`/blogs/count/${id}`);
-
-export const getBlogById = async ({ id }) => {
+export const getBlogById = async ({ slug, lang }) => {
     try {
-        const res = await axiosClient.get(`/blogs/${id}`);
+        const res = await axiosClient.get(`/blogs/${slug}`, {
+            params: { lang },
+        });
+
+        return res?.data;
+    } catch (err) {
+        toast.error(i18n.t("notify.fail"));
+    }
+};
+
+export const getBlogAdmin = async (params) => axiosClient.get("/admin/blogs", { params }).then((res) => res?.data);
+
+export const getBlogAdminById = async ({ id }) => {
+    try {
+        const res = await axiosClient.get(`/admin/blogs/${id}`);
 
         return res?.data;
     } catch (err) {
